Refactor styles: rename utilStyles to utils and migrate styles to new file
