+++ src/pages/Cards.tsx
@@ -1,7 +1,7 @@
 import { footballTeam } from "../data/footballTeam";
 import { FootballPlayer } from "../types/FootballPlayer";
 // import { TableHeader } from "../types/TableHeader";
-import utilStyles from "../styles/utilStyles.module.css";
+import utilStyles from "../styles/utils.module.css";
 import cardStyles from "../styles/cardStyles.module.css";
 import { useNavigate } from "react-router-dom";
