+++ src/pages/FantasyFootballTeam.tsx
@@ -1,8 +1,7 @@
 import { footballTeam } from "../data/footballTeam";
 import { FootballPlayer } from "../types/FootballPlayer";
 import { TableHeader } from "../types/TableHeader";
-import utilStyles from "../styles/utilStyles.module.css";
-
+import utilStyles from "../styles/utils.module.css";
 
 const playersList: FootballPlayer[] = footballTeam;
